[ Mixed ] fix apply using casted function
Describe a commit content (Until 80 colums per line) in detail ASAP.

**Changes proposed in this PR:**
- Added TOC generator for README.md

Resolves:

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

+++ test/unittest/unittest_nntrainer_tensor_fp16.cpp
@@ -3752,7 +3752,7 @@ TEST(nntrainer_Tensor, average_axis_p) {
   nntrainer::Tensor t = constant(1.0, 2, 2, 2, 2, nntrainer::Tformat::NCHW,
                                  nntrainer::Tdatatype::FP16);
   int idx = 0;
-  std::function<_Float16(_Float16)> f = [&](_Float16 in) { return idx++ % 2; };
+  std::function<float(float)> f = [&](float in) { return idx++ % 2; };
   t = t.apply(f);
 
   nntrainer::Tensor actual, expected;
